Import is_file_url from a module with no index client, 16% off nab config list
